+++ CHANGELOG.md
@@ -32,6 +32,7 @@ FEATURES:
   * `jwe:pbes2-a256kw:<ascii>` to initialize a PBES2 key derivation function for AES256 key wrapping with AES256 GCM Encryption transformer
 * sdk/transformer: Encryption transformer dynamic factory. [#80](https://github.com/elastic/harp/pull/80)
   * Use `github.com/elastic/harp/pkg/value/encryption.Register(prefix, factory)` to register a transformer factory matching the given prefix.
+* bundle/prefixer: parameter `--remove` added to support prefix removal operation [#81](https://github.com/elastic/harp/pull/81)
 
 CHANGES: